requirements.txt
fastapi==0.95.0
uvicorn>=0.22.0,<0.28.0
sqlalchemy>=2.0.20,<2.1.0
pydantic==1.10.12
python-multipart>=0.0.6,<0.1.0
python-dotenv>=1.0.0,<2.0.0
cloudinary>=1.37.0,<2.0.0 